changed param type onIRC31Received test

+++ core-contracts/StakedLP/src/test/java/finance/omm/score/test/unit/stakedLP/StakedLPTest.java
@@ -413,7 +413,7 @@ public class StakedLPTest extends AbstractStakedLPTest {
                 any(BigInteger.class));
 
         doNothing().when(scoreSpy).call(eq(Contracts.REWARDS),eq("handleLPAction"),any(),any());
-        score.invoke(account,"onIRC31Received", operator,from,id,amount,data);
+        score.invoke(account,"onIRC31Received", operator,from,BigInteger.valueOf(id),amount,data);
     }
